Update babacloud_api_analysis.py
I added a small value, "epsilon," to the monitor callback as sometimes it can happen that time_now is equal to _time_prev, and this results in a division by zero error. Adding a very small value does not change the speed variable significantly but avoid the error. Alternatively, you can add `if time_now - _time_prev == 0`, but I think this is a more elegant solution. Your choice. :)